move datagrid components to proper folder
